实现了主页文章列表隔行封面图效果，文章概要，使用io读取和创建md文件，查看文章等内容

+++ src/main/java/com/wei/dao/UserDao.java
@@ -30,5 +30,5 @@ public interface UserDao extends BaseMapper<User> {
      * @return 用户信息
      */
     @Select("select u.uid,username,u.joindate,TIMESTAMPDIFF(DAY,u.joindate,now()) regtime,u.nickname,u.eamil,u.signature,u.phonenumber,u.tximg,count(a.uid) acount from article a join user u on a.uid = u.uid where u.uid=#{uid}")
-    UserVo searchUserInfo(Integer uid);
+    UserVo searchUserInfo(long uid);
 }
